+++ physical/physical.go
@@ -119,6 +119,7 @@ var builtinBackends = map[string]Factory{
 	"etcd":       newEtcdBackend,
 	"mysql":      newMySQLBackend,
 	"postgresql": newPostgreSQLBackend,
+	"swift":      newSwiftBackend,
 }
 
 // PermitPool is a wrapper around a semaphore library to keep things