+++ CHANGELOG.md
@@ -57,6 +57,8 @@ FEATURES:
    secret distribution easier and more secure, including secure introduction.
  * **Azure Physical Backend**: You can now use Azure blob object storage as
    your Vault physical data store [GH-1266]
+ * **Swift Physical Backend**: You can now use Swift blob object storage as
+   your Vault physical data store [GH-1425]
  * **Consul Backend Health Checks**: The Consul backend will automatically
    register a `vault` service and perform its own health checking. By default
    the active node can be found at `active.vault.service.consul` and all with
@@ -102,6 +104,9 @@ IMPROVEMENTS:
    favor of normal ACL mechanisms [GH-1312]
  * credential/token: Sanitize policies and add `default` policies in appropriate
    places [GH-1235]
+ * credential/token: Setting the renewable status of a token is now possible
+   via `vault token-create` and the API. The default is true, but tokens can be
+   specified as non-renewable. [GH-1499]
  * secret/aws: Use chain credentials to allow environment/EC2 instance/shared
    providers [GH-307]
  * secret/aws: Support for STS AssumeRole functionality [GH-1318]